refactor steps.py - split to more files
First step in refactoring the steps.py file. No functional change made,
but code is more readable, steps are grouped to following files:
 * bond_team_bridge - not so much of them to create separate files
 * commands - various commands like Execute, Ping, Wait...
 * connection - management of connections
 * device - management of various devices (nmcli device)
 * editor - editor related steps
 * files - file management steps
 * prepare - (long) prepare steps (usually firsts steps in scenarios)
           - maybe it would be better to convert these steps to prepare
scripts
 * service - controling state of NM (Start, Stop, Kill...)
 * vpn - vpn related steps
